requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
cloudinary==1.41.0
colorama==0.4.6
Django==5.1.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg2==2.9.9
PyJWT==2.9.0
pytest==8.3.3
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.3



